+++ src/_quixote_test.js
@@ -8,11 +8,11 @@ var Frame = require("./frame.js");
 describe("Quixote", function() {
 
 	it("creates frame", function(done) {
-		var frame = quixote.createFrame(600, 400, { src: "/base/src/_frame_test.html" }, function(callbackFrame) {
+		var frame = quixote.createFrame(600, 400, { src: "/base/src/_frame_test.html" }, function(err, callbackFrame) {
 			assert.noException(function() {
 				callbackFrame.getElement("#exists");
 			});
-			done();
+			done(err);
 		});
 		assert.type(frame, Frame, "createFrame() returns frame object immediately");
 	});